Option to include an EFI image into ISO images
--efi-boot option to include an EFI boot image as an alternative El
Torrito boot image into the ISO image. This is necessary for booting an
ISO image on most EFI implementations. Booting the ISO image also works
from USB devices when writeing the image directly to the device. The EFI
boot files must be present in binary/efi/boot already. Use a binary hook
or include to copy the files to this directory.

This code is an adaption of similar code present in debian-cd.